fix: [#1274] Fix live update
